update system reboot and shutdown

+++ api/system/setting/reboot.go
@@ -1,16 +1,27 @@
 package setting
 
 import (
+	"net/http"
 	"os/exec"
 
 	"github.com/gin-gonic/gin"
 	"github.com/open-cmi/cmmns/module/auditlog"
 	"github.com/open-cmi/cmmns/module/rbac"
 	"github.com/open-cmi/gobase/essential/i18n"
+	"github.com/open-cmi/gobase/pkg/goparam"
 )
 
 func Reboot(c *gin.Context) {
 	ah := auditlog.NewAuditHandler(c)
+
+	userparam := goparam.GetUser(c)
+	role := userparam["role"].(string)
+	if role != "admin" {
+		ah.InsertOperationLog(i18n.Sprintf("reboot system"), false)
+		c.JSON(http.StatusForbidden, "")
+		return
+	}
+
 	ah.InsertOperationLog(i18n.Sprintf("reboot system"), true)
 	exec.Command("/bin/sh", "-c", "reboot -f").Output()
 	c.JSON(200, gin.H{
@@ -21,8 +32,15 @@ func Reboot(c *gin.Context) {
 
 func ShutDown(c *gin.Context) {
 	ah := auditlog.NewAuditHandler(c)
-	ah.InsertOperationLog(i18n.Sprintf("shutdown system"), true)
+	userparam := goparam.GetUser(c)
+	role := userparam["role"].(string)
+	if role != "admin" {
+		ah.InsertOperationLog(i18n.Sprintf("shutdown system"), false)
+		c.JSON(http.StatusForbidden, "")
+		return
+	}
 
+	ah.InsertOperationLog(i18n.Sprintf("shutdown system"), true)
 	exec.Command("/bin/sh", "-c", "shutdown -h now").Output()
 	c.JSON(200, gin.H{
 		"ret": 0,
